+++ TwcRA002Tests.cs
@@ -1,19 +1,19 @@
-﻿using OpenQA.Selenium;
+﻿using OfficeOpenXml;
+using OpenQA.Selenium;
 using OpenQA.Selenium.Interactions;
 using OpenQA.Selenium.Support.UI;
 using SeleniumExtras.WaitHelpers;
 using System.Net;
-using OfficeOpenXml;
 using static NUnit.Framework.Assert;
 
 namespace DomainStorm.Project.TWC.Tests
 {
-    public class TwcRA001Tests
+    public class TwcRA002Tests
     {
         private IWebDriver _driver = null!;
         private WebDriverWait _wait = null!;
         private Actions _actions = null!;
-        public TwcRA001Tests()
+        public TwcRA002Tests()
         {
             TestHelper.CleanDb();
         }
@@ -34,7 +34,7 @@ namespace DomainStorm.Project.TWC.Tests
 
         [Test]
         [Order(0)]
-        public async Task TwcRA001_01()
+        public async Task TwcRA002_01()
         {
             //0511 建立表單
             TestHelper.AccessToken = await TestHelper.GetAccessToken();
@@ -115,39 +115,79 @@ namespace DomainStorm.Project.TWC.Tests
             _wait.Until(ExpectedConditions.UrlContains(targetUrl));
             TestHelper.ClickRow(_driver, TestHelper.ApplyCaseNo!);
 
+            logout = _wait.Until(ExpectedConditions.ElementIsVisible(By.CssSelector("storm-tooltip > div > a[href='./logout']")));
+            _actions.MoveToElement(logout).Click().Perform();
+
+            //ning53 建立表單
+            TestHelper.AccessToken = await TestHelper.GetAccessToken();
+            statusCode = await TestHelper.CreateForm(TestHelper.AccessToken!, $"{TestHelper.BaseUrl}/api/v1/bmTransferApply/confirm", Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Assets/twcweb-RA001_bmTransferApply.json"));
+
+            await TestHelper.Login(_driver, "ning53", TestHelper.Password!);
+            _driver.Navigate().GoToUrl($@"{TestHelper.BaseUrl}/draft");
+            TestHelper.ClickRow(_driver, TestHelper.ApplyCaseNo!);
+
+            _wait.Until(ExpectedConditions.ElementExists(By.CssSelector("iframe")));
+            _driver.SwitchTo().Frame(0);
+
+            acceptSign = _wait.Until(ExpectedConditions.ElementExists(By.CssSelector("[id='受理'] > span")));
+            ((IJavaScriptExecutor)_driver).ExecuteScript("arguments[0].click();", acceptSign);
+
+            _driver.SwitchTo().DefaultContent();
+
+            addFileButton = _wait.Until(ExpectedConditions.ElementExists(By.CssSelector("[id='file'] > div.float-end > button")));
+            _actions.MoveToElement(addFileButton).Click().Perform();
+
+            lastHiddenInput = _wait.Until(ExpectedConditions.ElementExists(By.CssSelector("input.dz-hidden-input:nth-of-type(3)")));
+            twcweb_01_1_夾帶附件1 = "twcweb_01_1_夾帶附件1.pdf";
+            filePath = Path.Combine(Directory.GetCurrentDirectory(), "Assets", twcweb_01_1_夾帶附件1);
+            lastHiddenInput.SendKeys(filePath);
+
+            uploadButton = _wait.Until(ExpectedConditions.ElementIsVisible(By.CssSelector("div.d-flex.justify-content-end.mt-4 button[name='button']")));
+            _actions.MoveToElement(uploadButton).Click().Perform();
+            That(TestHelper.WaitStormEditTableUpload(_driver, "storm-table-cell > span")!.Text, Is.EqualTo("twcweb_01_1_夾帶附件1.pdf"));
+
+            checkButton = _wait.Until(ExpectedConditions.ElementExists(By.CssSelector("[id='用印或代送件只需夾帶附件']")));
+            _actions.MoveToElement(checkButton).Click().Perform();
+
+            infoButton = _wait.Until(ExpectedConditions.ElementExists(By.CssSelector("button.btn.bg-gradient-info.m-0.ms-2")));
+            _actions.MoveToElement(infoButton).Click().Perform();
+
+            targetUrl = $"{TestHelper.BaseUrl}/unfinished";
+            _wait.Until(ExpectedConditions.UrlContains(targetUrl));
+            TestHelper.ClickRow(_driver, TestHelper.ApplyCaseNo!);
+
             var signNumber = TestHelper.FindAndMoveElement(_driver, "storm-card:nth-child(9) > div.row > div.col-sm-7");
             That(signNumber.GetAttribute("textContent"), Is.EqualTo(TestHelper.ApplyCaseNo));
         }
 
         [Test]
         [Order(1)]
-        public async Task TwcRA001_02To09()
+        public async Task TwcRA002_02To09()
         {
-            await TwcRA001_02();
-            await TwcRA001_03();
-            await TwcRA001_04();
-            await TwcRA001_05();
-            await TwcRA001_06();
-            await TwcRA001_07();
-            await TwcRA001_08();
-            await TwcRA001_09();
+            await TwcRA002_02();
+            await TwcRA002_03();
+            await TwcRA002_04();
+            await TwcRA002_05();
+            await TwcRA002_06();
+            await TwcRA002_07();
+            await TwcRA002_08();
+            await TwcRA002_09();
         }
-        public async Task TwcRA001_02()
+        public async Task TwcRA002_02()
         {
             TestHelper.AccessToken = await TestHelper.GetAccessToken();
             That(TestHelper.AccessToken, Is.Not.Empty);
         }
 
-        public async Task TwcRA001_03()
-
+        public async Task TwcRA002_03()
         {
-            HttpStatusCode statusCode = await TestHelper.CreateForm(TestHelper.AccessToken!, $"{TestHelper.BaseUrl}/api/v1/bmTransferApply/confirm", Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Assets/twcweb-RA001_bmTransferApply.json"));
+            HttpStatusCode statusCode = await TestHelper.CreateForm(TestHelper.AccessToken!, $"{TestHelper.BaseUrl}/api/v1/bmTransferApply/confirm", Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Assets/twcweb-RA002_bmTransferApply.json"));
             That(statusCode, Is.EqualTo(HttpStatusCode.OK));
         }
 
-        public async Task TwcRA001_04()
+        public async Task TwcRA002_04()
         {
-            await TestHelper.Login(_driver, "ning53", TestHelper.Password!);
+            await TestHelper.Login(_driver, "0511", TestHelper.Password!);
             _driver.Navigate().GoToUrl($@"{TestHelper.BaseUrl}/draft");
             TestHelper.ClickRow(_driver, TestHelper.ApplyCaseNo!);
 
@@ -155,7 +195,7 @@ namespace DomainStorm.Project.TWC.Tests
             That(pTitle!.Text, Is.EqualTo("沒有找到符合的結果"));
         }
 
-        public async Task TwcRA001_05()
+        public async Task TwcRA002_05()
         {
             var addFileButton = _wait.Until(ExpectedConditions.ElementExists(By.CssSelector("[id='file'] > div.float-end > button")));
             _actions.MoveToElement(addFileButton).Click().Perform();
@@ -169,14 +209,14 @@ namespace DomainStorm.Project.TWC.Tests
             That(fileName.GetAttribute("value"), Is.EqualTo("twcweb_01_1_夾帶附件1.pdf"));
         }
 
-        public async Task TwcRA001_06()
+        public async Task TwcRA002_06()
         {
             var uploadButton = _wait.Until(ExpectedConditions.ElementIsVisible(By.CssSelector("div.d-flex.justify-content-end.mt-4 button[name='button']")));
             _actions.MoveToElement(uploadButton).Click().Perform();
             That(TestHelper.WaitStormEditTableUpload(_driver, "storm-table-cell > span")!.Text, Is.EqualTo("twcweb_01_1_夾帶附件1.pdf"));
         }
 
-        public async Task TwcRA001_07()
+        public async Task TwcRA002_07()
         {
             _driver.SwitchTo().Frame(0);
 
@@ -188,7 +228,7 @@ namespace DomainStorm.Project.TWC.Tests
             That(chceckSign!, Is.Not.Null);
         }
 
-        public async Task TwcRA001_08()
+        public async Task TwcRA002_08()
         {
             _driver.SwitchTo().DefaultContent();
 
@@ -197,7 +237,7 @@ namespace DomainStorm.Project.TWC.Tests
             That(checkButton.GetAttribute("checked"), Is.EqualTo("true"));
         }
 
-        public async Task TwcRA001_09()
+        public async Task TwcRA002_09()
         {
             var infoButton = _wait.Until(ExpectedConditions.ElementExists(By.CssSelector("button.btn.bg-gradient-info.m-0.ms-2")));
             _actions.MoveToElement(infoButton).Click().Perform();
@@ -212,69 +252,48 @@ namespace DomainStorm.Project.TWC.Tests
 
         [Test]
         [Order(2)]
-        public async Task TwcRA001_10To11()
-        {
-            await TwcRA001_10();
-            await TwcRA001_11();
-        }
-        public async Task TwcRA001_10()
+        public async Task TwcRA002_10()
         {
             await TestHelper.Login(_driver, "0511", TestHelper.Password!);
-        }
+            _driver.Navigate().GoToUrl($@"{TestHelper.BaseUrl}/report/RA002");
 
-        public async Task TwcRA001_11()
-        {
-            _driver.Navigate().GoToUrl($@"{TestHelper.BaseUrl}/report/RA001");
-            _wait.Until(ExpectedConditions.ElementExists(By.CssSelector("iframe")));
+            _wait.Until(ExpectedConditions.ElementIsVisible(By.CssSelector("iframe")));
+
             _driver.SwitchTo().Frame(0);
 
-            var 區處別 = _wait.Until(ExpectedConditions.ElementToBeClickable(By.CssSelector("storm-card > form > div.mt-3 > storm-select >div.choices")));
+            var 區處別 = _wait.Until(ExpectedConditions.ElementExists(By.CssSelector("storm-card storm-select div.choices")));
             _actions.MoveToElement(區處別).Click().Perform();
 
             var 第四區管理處 = _wait.Until(ExpectedConditions.ElementExists(By.CssSelector("div.choices__list.choices__list--dropdown > div.choices__list > [data-id='2']")));
             _actions.MoveToElement(第四區管理處).Click().Perform();
 
-            var applyDateBegin = _wait.Until(ExpectedConditions.ElementExists(By.CssSelector("[label='受理日期起']")));
-            var input = applyDateBegin.GetShadowRoot().FindElement(By.CssSelector("input"));
-            _actions.MoveToElement(applyDateBegin).Click().Perform();
+            var 選擇年份 = _wait.Until(ExpectedConditions.ElementExists(By.CssSelector("[label='選擇年份'] div.choices")));
+            _actions.MoveToElement(選擇年份).Click().Perform();
 
-            var select = _wait.Until(ExpectedConditions.ElementExists(By.CssSelector("div.flatpickr-calendar.open div.flatpickr-current-month select")));
-            var applyMonthBegin = new SelectElement(select);
-            applyMonthBegin.SelectByText("March");
-
-            var applyDayBegin = _wait.Until(ExpectedConditions.ElementExists(By.CssSelector("div.flatpickr-calendar.open div.flatpickr-innerContainer div.flatpickr-days span[aria-label='March 6, 2023']")));
-            _actions.MoveToElement(applyDayBegin).Click().Perform();
-
-            var applyDateEnd = _wait.Until(ExpectedConditions.ElementExists(By.CssSelector("[label='受理日期迄']")));
-            input = applyDateEnd.GetShadowRoot().FindElement(By.CssSelector("input"));
-            _actions.MoveToElement(applyDateEnd).Click().Perform();
-
-            select = _wait.Until(ExpectedConditions.ElementExists(By.CssSelector("div.flatpickr-calendar.open div.flatpickr-current-month select")));
-            var applyMonthEnd = new SelectElement(select);
-            applyMonthEnd.SelectByText("April");
-
-            var applyDatEnd = _wait.Until(ExpectedConditions.ElementExists(By.CssSelector("div.flatpickr-calendar.open div.flatpickr-innerContainer div.flatpickr-days span[aria-label='April 6, 2023']")));
-            _actions.MoveToElement(applyDatEnd).Click().Perform();
+            var 年份 = _wait.Until(ExpectedConditions.ElementExists(By.CssSelector("div.choices__list [data-value='2023']")));
+            _actions.MoveToElement(年份).Click().Perform();
 
             var 檔案格式 = _wait.Until(ExpectedConditions.ElementExists(By.CssSelector("[label='檔案格式'] div.choices")));
             _actions.MoveToElement(檔案格式).Click().Perform();
 
-            var xlsx = _wait.Until(ExpectedConditions.ElementIsVisible(By.CssSelector("div.choices__list [data-value='XLSX']")));
+            var xlsx = _wait.Until(ExpectedConditions.ElementExists(By.CssSelector("div.choices__list [data-value='XLSX']")));
             _actions.MoveToElement(xlsx).Click().Perform();
 
-            That(TestHelper.DownloadFileAndVerify(_driver, "RA001.xlsx", "storm-card > form > div > button"), Is.True);
+            That(TestHelper.DownloadFileAndVerify(_driver, "RA002.xlsx", "storm-card > form > div > button"), Is.True);
 
             string _downloadDirectory = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.UserProfile), "Downloads");
-            string filePath = Path.Combine(_downloadDirectory, "RA001.xlsx");
+            string filePath = Path.Combine(_downloadDirectory, "RA002.xlsx");
 
             ExcelPackage.LicenseContext = LicenseContext.NonCommercial;
+
             using var package = new ExcelPackage(new FileInfo(filePath));
             var worksheet = package.Workbook.Worksheets[0];
 
+            string D6Value = worksheet.Cells["D6"].Text;
+            string F6Value = worksheet.Cells["F6"].Text;
             string E7Value = worksheet.Cells["E7"].Text;
-            string E8Value = worksheet.Cells["E8"].Text;
 
-            That(E7Value == "2" && E8Value == "1", Is.True);
+            That(E7Value == "1" && D6Value == "2" && F6Value == "1", Is.True);
         }
     }
 }
